requirements.txt
antlr4-python3-runtime==4.5.3
click==6.7
decorator==4.0.10
Flask==0.12
Flask-JSON==0.3.2
itsdangerous==0.24
Jinja2==2.9.4
MarkupSafe==0.23
python-dateutil==2.6.0
self==0.0.5
six==1.10.0
SQLAlchemy==1.1.5
Werkzeug==0.11.15
